Fix automatic update of class diagram in README.md
This class diagram is generated automatically from the classDiagram.puml file whenever it is updated. 
The classDiagram.puml file is in plantUML format (ref. http://plantuml.com/class-diagram) .
In order for this to work I had to add the "cache=no" parameter to the URL otherwise the shown image would be a cached version and not reflecting updates to the classDiagram.puml file.

Note: Still not sure this will work. Future updates of the classDiagram.puml file will tell us.